fix: PersistenceManager:default can't be resolved e2e tests resolves #47
